requirements.txt
# Core ETL dependencies
lxml>=4.9.0
python-dateutil>=2.8.2

# Database
# SQLite is built into Python, no additional package needed

# Optional API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Development
pytest>=7.4.0
